Abarelli
Added the Fuseki Server with a built in basic /Test directory, which autoloads when the server is started.

To run the server use: ./fuseki-server --update

To run the server and bootup the Demo directory, use: ./fuseki-server --update --file=Demo/BuildTest.rdf /Demo

Currently the Demo works with rdf, rdfs, and skos.

Next step: Add our own ontology and get the Demo setup to reference multiple files.